requirements.txt
phonenumbers
web-browser
exifread
requests
googlesearch-python
